+++ epic_app/tests/externals/ERAMVisuals/test_eram_visuals_wrapper.py
@@ -1,42 +1,60 @@
-from epic_app.externals.ERAMVisuals.eram_visuals_wrapper import EramVisualsWrapper
-from epic_app.externals.external_wrapper_base import ExternalWrapperStatusType
+import shutil
+from unittest.mock import Mock
+
+import pytest
+
+from epic_app.externals.ERAMVisuals.eram_visuals_wrapper import (
+    EramVisualsRunner,
+    EramVisualsWrapper,
+)
+from epic_app.externals.external_wrapper_base import (
+    ExternalRunner,
+    ExternalWrapperStatusType,
+)
 from epic_app.tests import test_data_dir
 
 
 class TestEramVisualsWrapper:
-    def test_execute_r_snippet_with_sample_data_succeeds(self):
+    def test_execute_r_snippet_with_sample_data_succeeds(
+        self, request: pytest.FixtureRequest
+    ):
         # 1. Define test data.
         _csv_file = test_data_dir / "csv" / "evo_summary.csv"
         assert _csv_file.exists()
-        _png_file = test_data_dir / "test_evo_summary.png"
-        if _png_file.is_file():
-            _png_file.unlink()
-        if _png_file.with_suffix(".pdf").is_file():
-            _png_file.with_suffix(".pdf").unlink()
+        _output_dir = test_data_dir / request.node.name
+        if _output_dir.exists():
+            shutil.rmtree(_output_dir)
 
         # 2. Run test.
-        eram_visuals = EramVisualsWrapper(input_file=_csv_file, output_file=_png_file)
+        eram_visuals = EramVisualsWrapper(
+            input_file=_csv_file, output_dir=_output_dir, runner=EramVisualsRunner
+        )
         eram_visuals.execute()
 
         # 3. Verify final expectations.
         assert eram_visuals.status.status_type == ExternalWrapperStatusType.SUCCEEDED
-        assert _png_file.exists()
-        assert _png_file.with_suffix(".pdf").exists()
+        assert eram_visuals.output
+        assert eram_visuals.output.png_output.exists()
+        assert eram_visuals.output.pdf_output.exists()
 
-    def test_given_failed_execute_status_is_failed(self):
+    def test_given_failed_execute_status_is_failed(
+        self, request: pytest.FixtureRequest
+    ):
         _exception_mssg = "Ea aliqua culpa occaecat minim reprehenderit et."
 
-        class MockEramVisualsWrapper(EramVisualsWrapper):
-            def _run_script(self) -> None:
+        class MockEramRunner(ExternalRunner):
+            def run(self, *args, **kwargs) -> None:
                 raise Exception(_exception_mssg)
 
         # 1. Define test data.
+        _output_dir = test_data_dir / request.node.name
+        if _output_dir.exists():
+            shutil.rmtree(_output_dir)
         _csv_file = test_data_dir / "csv" / "evo_summary.csv"
-        _png_file = test_data_dir / "test_failing_evo_summary.png"
 
         # 2. Run mocked up test
-        _test_wrapper = MockEramVisualsWrapper(
-            input_file=_csv_file, output_file=_png_file
+        _test_wrapper = EramVisualsWrapper(
+            input_file=_csv_file, output_dir=_output_dir, runner=MockEramRunner
         )
         _test_wrapper.execute()
 
@@ -44,5 +62,6 @@ class TestEramVisualsWrapper:
         assert _test_wrapper.status.status_type == ExternalWrapperStatusType.FAILED
         assert _test_wrapper.status.status_info == _exception_mssg
         assert str(_test_wrapper.status) == f"Failed: {_exception_mssg}"
-        assert not _png_file.exists()
-        assert not _png_file.with_suffix(".pdf").exists()
+        assert _test_wrapper.output
+        assert not _test_wrapper.output.pdf_output.exists()
+        assert not _test_wrapper.output.png_output.exists()